refactor: copilot suggestions
